+++ pkg/httpclient/client.go
@@ -11,7 +11,7 @@ import (
 
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 
-	"github.com/docker/docker-agent/pkg/remote"
+	"github.com/docker/docker-agent/pkg/desktop/transport"
 	"github.com/docker/docker-agent/pkg/userid"
 	"github.com/docker/docker-agent/pkg/version"
 )
@@ -173,8 +173,8 @@ func WithQuery(query url.Values) Opt {
 
 // newTransport returns an HTTP transport with automatic gzip compression disabled and using Docker Desktop proxy if available.
 func newTransport(ctx context.Context) http.RoundTripper {
-	// Get the base transport with Desktop proxy support from remote package
-	rt := remote.NewTransport(ctx)
+	// Get the base transport with Desktop proxy support
+	rt := transport.New(ctx)
 
 	// Disable compression for SSE streaming compatibility
 	// Handle both direct *http.Transport and the fallback transport wrapper
